benchmarks/requirements.txt
seaborn==0.13.2
python-dateutil==2.8.2
matplotlib==3.8.2
